refactor: some messaged fixed and some unnecessary spaces removed

+++ src/main/java/gustavo/ventieri/capitalmind/application/dto/cryptoCurrency/CryptoCurrencyResponseDto.java
@@ -1,5 +1,5 @@
 package gustavo.ventieri.capitalmind.application.dto.cryptoCurrency;
 
-public record CryptoCurrencyResponseDto(Long stockId, String name, String description, Integer quantity, Double price) {
+public record CryptoCurrencyResponseDto(Long stockId, String name, String description, Double quantity, Double price) {
     
 }
